Handle inserts that have a newIndexPath
Work around a bug in Xcode 7.0 and 7.1 when running on iOS 8 - updated objects
sometimes result in both an Update *and* and Insert call to didChangeObject, which
makes no sense. Thankfully the bad Inserts have a non-nil "old" indexPath (which
also makes no sense) - we check for that here and ignore those erroneous messages.

For more discussion, see https://forums.developer.apple.com/thread/12184